Backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pydantic[email]
yt-dlp
requests
supabase
openai
python-dotenv